Move blob_db/ttl_extractor.h into blob_db/blob_db.h
